ocr_service/requirements.txt
torch>=2.0.0
transformers>=4.46.3
tokenizers>=0.20.3
einops
addict
easydict
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=1.24.0
requests>=2.31.0
matplotlib>=3.7.0
torchvision>=0.15.0


